refactor(dashboard): widgets folder location
