+++ WebApp/frontend/src/components/app/views/Dashboard/widgets/HistoryWidget.tsx
