+++ WebApp/frontend/src/components/app/views/Dashboard/widgets/StatisticsWidget.tsx
