+++ WebApp/frontend/src/components/app/views/Dashboard/widgets/TranscriptWidget.tsx
